feat:
1.新增客户端主动向服务端注册/下线映射
2.管理面板增加映射来源字段
3.增加SpringBoot starter用于集成etp
4.配置初始化功能完善
5.通信协议优化

+++ etp-server/src/main/java/com/xiaoniucode/etp/server/web/SQLiteUtils.java
@@ -93,7 +93,7 @@ public final class SQLiteUtils {
         return array;
     }
 
-    public static long insert(String sql, Object... params) {
+    public static int insert(String sql, Object... params) {
         Connection conn = null;
         try {
             conn = getConnection();
@@ -110,7 +110,7 @@ public final class SQLiteUtils {
 
                 try (ResultSet rs = pstmt.getGeneratedKeys()) {
                     if (rs.next()) {
-                        long id = rs.getLong(1);
+                        int id = rs.getInt(1);
                         logger.debug("插入成功，ID = {}", id);
                         return id;
                     } else {
